requirements.txt
amqp==5.2.0
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==2.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses==0.8
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.1.0
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pkg-resources==0.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.8
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.2
redis==4.3.6
six==1.16.0
SQLAlchemy==1.4.54
typing-extensions==4.1.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.0.3
zipp==3.6.0
